backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
bcrypt
redis
celery
stripe
google-genai
slowapi
python-multipart
httpx
email-validator
